Ongoing development

+++ OMGBlockz/trunk/src/com/cybertivity/omgblockz/dimensions/Nether.java
@@ -4,9 +4,9 @@ import com.cybertivity.omgblockz.utility.FileSystemHelper;
 
 public class Nether extends DimensionBase {
 
-    private static final int WORLD_HEIGHT = 256;
-    private static final int MAX_LAND_HEIGHT = 256;
-    private static final int SEA_LEVEL = -1;
+    private static final short WORLD_HEIGHT = 256;
+    private static final short MAX_LAND_HEIGHT = 256;
+    private static final short SEA_LEVEL = -1;
     
     public Nether(int seed, String path)  {
         super(Dimension.NETHER, seed,
