+++ OMGBlockz/trunk/src/com/cybertivity/omgblockz/dimensions/Overworld.java
@@ -7,9 +7,9 @@ import java.util.ArrayList;
 
 public class Overworld extends DimensionBase {
 
-    private static final int WORLD_HEIGHT = 256;
-    private static final int MAX_LAND_HEIGHT = 256;
-    private static final int SEA_LEVEL = 64;
+    private static final short WORLD_HEIGHT = 256;
+    private static final short MAX_LAND_HEIGHT = 256;
+    private static final short SEA_LEVEL = 64;
 
     static {
         biomes.add(new Plains());
